customer-portal/pyproject.toml
[project]
name = "customer-portal-api"
version = "0.1.0"
description = "SUNET OpenStack Customer Self-Service Portal API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "gitpython>=3.1.0",
    "httpx>=0.28.0",
    "authlib>=1.4.0",
    "itsdangerous>=2.2.0",
    "pyyaml>=6.0",
    "kubernetes>=31.0.0",
    "openstacksdk>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
